fix: use i32 instead of u8 for CborMapValue
This way, the data type actually corresponds to
the specification in the IANA registry.

+++ src/model/cbor_values.rs
@@ -16,9 +16,9 @@ type ByteStringValue = Vec<u8>;
 #[derive(Debug, Deserialize, PartialEq, Eq, Default)]
 pub struct ByteString(ByteStringValue);
 
-pub struct CborMapValue<T>(pub T) where u8: Into<T>, T: Into<u8> + Copy;
+pub struct CborMapValue<T>(pub T) where i32: Into<T>, T: Into<i32> + Copy;
 
-impl<T> Deref for CborMapValue<T> where T: From<u8> + Into<u8> + Copy
+impl<T> Deref for CborMapValue<T> where T: From<i32> + Into<i32> + Copy
 {
     type Target = T;
 
@@ -27,20 +27,20 @@ impl<T> Deref for CborMapValue<T> where T: From<u8> + Into<u8> + Copy
     }
 }
 
-impl<T> Serialize for CborMapValue<T> where T: From<u8> + Into<u8> + Copy {
+impl<T> Serialize for CborMapValue<T> where T: From<i32> + Into<i32> + Copy {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error> where S: Serializer {
-        let cbor_value: u8 = self.0.into();
+        let cbor_value: i32 = self.0.into();
         Value::from(cbor_value).serialize(serializer)
     }
 }
 
-impl<'de, T> Deserialize<'de> for CborMapValue<T> where T: From<u8> + Into<u8> + Copy {
+impl<'de, T> Deserialize<'de> for CborMapValue<T> where T: From<i32> + Into<i32> + Copy {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
         where
             D: Deserializer<'de>,
     {
         if let Ok(Value::Integer(i)) = Value::deserialize(deserializer) {
-            Ok(CborMapValue(u8::try_from(i)
+            Ok(CborMapValue(i32::try_from(i)
                 .map_err(|x| D::Error::custom(x.to_string()))?.into()))
         } else {
             Err(D::Error::custom("CBOR map value must be an Integer!"))
